fix clear easyjson files

+++ protoc-go.sh
@@ -84,10 +84,7 @@ wait
 printf "\e[32mDone \e[32m(%.1f sec)\e[m\n" $(echo "$(date +%s.%N) - $starttime" | bc)
 
 # GENERATEING JSON
-
-if [[ $i == *".proto" ]]; then
-	rm -rf */*easyjson.go
-fi
+[ "$1" = 'json' ] && rm -rf */*easyjson.go
 
 starttime=$(date +%s.%N)
 printf "\e[32mgenerating json... (this could take up to 20 sec)"
